+++ libs/langchain/pyproject.toml
@@ -129,7 +129,6 @@ markdownify = {version = "^0.11.6", optional = true}
 assemblyai = {version = "^0.17.0", optional = true}
 dashvector = {version = "^1.0.1", optional = true}
 sqlite-vss = {version = "^0.1.2", optional = true}
-vowpal-wabbit-next = {version = "0.6.0", optional = true}
 
 
 [tool.poetry.group.test.dependencies]
@@ -343,11 +342,9 @@ extended_testing = [
  "xmltodict",
  "faiss-cpu",
  "openapi-schema-pydantic",
- "sentence-transformers",
  "markdownify",
  "dashvector",
  "sqlite-vss",
- "vowpal-wabbit-next",
 ]
 
 [tool.ruff]
